Add files via upload
The “Sorting Visualizer” project is basically a web development project which is used to visualize the sorting algorithms and performs the operation of algorithms step by step as the algorithms works like Bubble Sort, Quick Sort , Merge Sort ,Heap Sort. We shuffle the bars in a random order that is in a unsorted order. We perform sorting algorithms on these bars to sort these bars in a ascending order by applying the different algorithms by clicking on the buttons of these sorting algorithms